# app/Http/Controllers/Contact/ContactController.php
<?php

namespace App\Http\Controllers\Contact;

use App\Http\Controllers\Controller;
use App\Models\Contact;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Validator;

class ContactController extends Controller
{
    /**
     * @OA\Get(
     *      path="/api/getAllContactData",
     *      operationId="getContact",
     *      tags={"Contact"},
     *      summary="Get",
     *      description="Returns list",
     *      @OA\Response(
     *          response=200,
     *          description="Successful",
     * *          @OA\MediaType(
     *           mediaType="application/json",
     *      )
     *       ),
     *     )
     */
    public function getContact()
    {
        $data = Contact::latest()->get();

        $result = [
            'message' => "success",
            'success' => true,
            'status' => 200,
            'data' => $data
        ];
        return response()->json($result);
    }

    /**
     * @OA\Post(
     * path="/api/createContact",
     * summary="Create",
     * description="Creation",
     * security={{ "bearerAuth":{ }}},
     * operationId="storeContact",
     * tags={"Contact"},
     * @OA\RequestBody(
     *    required=true,
     *    description="Enregistrer",
     *    @OA\JsonContent(
     *       required={"name","address","message","email","phone"},
     *       @OA\Property(property="name", type="string", format="text",example="winner kambale"),
     *       @OA\Property(property="address", type="string", format="text",example="rdc"),
     *       @OA\Property(property="message", type="string", format="text",example="mpesa"),
     *       @OA\Property(property="phone", type="string", format="text",example="0997604471"),
     *       @OA\Property(property="email", type="string", format="text", example="winner@gmail.com")
     *    ),
     * ),
     * @OA\Response(
     *    response=201,
     *    description="success",
     *     ),
     * @OA\Response(
     *    response=401,
     *    description="ysteme",
     *     )
     * )
     */
    public function storeContact(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'name' => 'required',
            'address' => 'required',
            'email' => 'required|email',
            'message' => 'required',
            'phone' => 'required'
        ]);
        if ($validator->fails()) {
            return response()->json([
                'message' => 'Les données envoyées ne sont pas valides.',
                'errors' => $validator->errors()
            ], 422);
        }
        Contact::create([
            'name' => $request->name,
            'address' => $request->address,
            'email' => $request->email,
            'message' => $request->message,
            'phone' => $request->phone
        ]);
        $result = [
            'message' => "success",
            'success' => true,
            'status' => 201
        ];
        return response()->json($result);
    }

    /**
     * @OA\Delete(
     * path="/api/deleteContact/{id}",
     * summary="Delete",
     * description="Suppression",
     * security={{ "bearerAuth":{ }}},
     * operationId="deleteContact",
     * tags={"Contact"},
     * @OA\Parameter(
     *    name="id",
     *    in="path",
     *    required=true,
     *    @OA\Schema(type="integer")
     * ),
     * @OA\Response(
     *    response=200,
     *    description="Project deleted successfully",
     * ),
     * @OA\Response(
     *    response=404,
     *    description="Project not found",
     * ),
     * @OA\Response(
     *    response=401,
     *    description="Unauthorized",
     * )
     * )
     */
    public function deleteContact($id)
    {
        $contact = Contact::find($id);

        if (!$contact) {
            return response()->json(['message' => 'not found'], 404);
        }
        $contact->delete();

        return response()->json([
            'message' => "successfully",
            'success' => true,
            'status' => 200
        ]);
    }
}