# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Website John Services Motel API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/createContact": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Create",
                "description": "Creation",
                "operationId": "storeContact",
                "requestBody": {
                    "description": "Enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "message",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "message": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "0997604471"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "ysteme"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deleteContact/{id}": {
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete",
                "description": "Suppression",
                "operationId": "deleteContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/createRestaurant": {
            "post": {
                "tags": [
                    "Restaurant"
                ],
                "summary": "Create",
                "description": "Creation",
                "operationId": "createRestaurant",
                "requestBody": {
                    "description": "Enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title_en",
                                    "title_fr",
                                    "description_en",
                                    "description_fr"
                                ],
                                "properties": {
                                    "title_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "title_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "description_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "description_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "ysteme"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/updateRestaurant/{id}": {
            "put": {
                "tags": [
                    "Restaurant"
                ],
                "summary": "Update",
                "description": "Modification",
                "operationId": "updateRestaurant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Modifier",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title_en",
                                    "title_fr",
                                    "description_fr",
                                    "description_en"
                                ],
                                "properties": {
                                    "description_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner services"
                                    },
                                    "description_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner services"
                                    },
                                    "title_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "title_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}